Fixed building and fixed rename of defaults
