--- src/barcodes/index.tmp.js
@@ -1,19 +1,0 @@
-import {CODE39} from './CODE39/';
-import {CODE128, CODE128A, CODE128B, CODE128C} from './CODE128/';
-import {EAN13, EAN8, EAN5, EAN2, UPC} from './EAN_UPC/';
-import {ITF14} from './ITF14/';
-import {ITF} from './ITF/';
-import {MSI, MSI10, MSI11, MSI1010, MSI1110} from './MSI/';
-import {pharmacode} from './pharmacode/';
-import {GenericBarcode} from './GenericBarcode/';
-
-export default {
-	CODE39,
-	CODE128, CODE128A, CODE128B, CODE128C,
-	EAN13, EAN8, EAN5, EAN2, UPC,
-	ITF14,
-	ITF,
-	MSI, MSI10, MSI11, MSI1010, MSI1110,
-	pharmacode,
-	GenericBarcode
-};
